Prevent scrollbars from body element.
In preparation for the app to become fullscreen (and atop that, adjustable), we're controlling the body's margin programatically (so that I can use the variable in a later calc).

We're also hiding overflow from the body in the ASSUMPTION that the site's intention is to fullscreen the app and have naught else accessible.