initial and basic resize function.
The resize function is not tied to any particular game logic to keep it precise on its task (i.e., the viewport doesn't come into account here, can be handled alongside game logic).

What this function is yet to do, however, is command the game to shrink in accordance with the canvas size.