Initial thoughts on additional utilities for handling Node-centric tasks.
Not to be confused with NodeJS, of course. AppNodes, rather, as will be typed in the following commit.

As explained in the comments, I chose to have NodeUtil consume an app to be activated in thinking ahead - I don't want NodeUtil to be reliant on app being ready/initialized before NodeUtil declared. This is an abundance of upfront work for long term safety.

NodeUtil becomes closer to a pure function, after all, as it's not reliant on a particular piece of state being ready.

In addition, I have the game automatically start adjusting the viewport - both once manually at start, and then anytime the browser window is resized.