Adding the base properties we'd see on a node, now called AppNode.
Added additional properties for a node that should be able to handle momentum-related functionality, called a MovingNode (jnspired by the 'direction' from the example).

While the paddles for ping-pong will only ever move one direction, I feel that the ball is under the same category, hence supplying support for 2D movement in MovingNode.