Collision Detection...
First, I define some types that I suspect will be common throughout the NodeUtil code. XYAdjust, representing desired change in xy coords; and CollisionReason, representing the types of reasons that a Collision was deemed to have occurred.

As for the function itself, it's incredibly straightforward. Check if collision has occurred on any proposed (prop_) position, taking into account the height/width of the object itself. clearRect draws from the top left corner, so that check only needs to be made at the far ends.

Then we independently check the X and Y to see who, if any or both, is at fault, and return those findings to our consumer.