Both boxes can now be controlled.
There is not an incredible lot to add here. We will be handling Game logic (which is separate from the engine - aka app - logic) in Game. Players are hardcoded because Pong is built for two alone.

One could argue that there is some DRY violation in the eventListeners, but the logic/work isn't yet complex enough to warrant shifting the code. If we add deacceleration, however, it certainly will be worth changing.