Node Collision tweaks...
As I built other functions, certain data necessary from here changed. Further, there was a logical fault in thinking that X or Y would suffice just as with normal detection. While that may work for boundaries, it didn't work here. It only makes sense that the collision must verify that the attempted move is within the box, not past it in 2d space!