Bounce Hard, or Node<=>Node bouncing.
This took many attempts and ideas. The closest I've managed to theorize a method for simulating the shift in the behavior from a ball is this, though it's almost too programmatic.

Without specific requirements to tackle this sort of calculation, what I did was assess where the ball was landing on the target - particularly, on which surface and how far away from the center of that surface. The further away from the center, the more angled the ball flies. The closer to the center, the straighter it is.

Originally was, as can be seen through the commented out code, attempting to grab the distances that x/y had to the respective ends of each layer, but it turned out to be too inexact. I instead just went ahead and computed both 'weights' (how far are you away from the middle on a scale of -100% to 100%?). The weight of the side the bounce didn't occur on is greater than ABS(100), so I just pick the lower weight. I strongly suspect there are still instances where the weight is over ABS(100) on both sides, but for the purposes of this demo with no SME in the field of pong physics, I've given my best shot.